pyproject.toml
[project]
name = "multi-query-rag"
version = "0.1.0"
description = "RAG system with multi-query support for medical reports"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.0",
    "openai>=1.86.0",
    "pandas>=2.3.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
]

[project.scripts]
rag-query = "multi_query_rag.main:main"
rag-seed = "multi_query_rag.seed:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
